# Optimize BFGS for spinel