Optimizes BFGS for the spinel case. I see a speed up of ~30/40% on my machine. 

Main branch:
```julia-repl
julia> @be null_min_BFGS!($gv,$ph,$gm)
Benchmark: 3173 samples with 22 evaluations
 min    1.291 μs
 median 1.305 μs
 mean   1.310 μs
 max    2.350 μs
```

This PR:
```julia-repl
julia> @be null_min_BFGS!($gv,$ph,$gm)
Benchmark: 2762 samples with 33 evaluations
 min    872.727 ns
 median 881.818 ns
 mean   884.147 ns
 max    1.621 μs
```